pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "JoyIT_st7920"
version = "1.0.1"
dependencies = ['gpiozero', 'spidev', 'Pillow']
authors = [
  { name="Joy-IT", email="service@joy-it.net" },
]
description = "Library to use the SBC-LCD128x64 with the Raspberry Pi"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/joy-it/JoyIT_st7920"
Issues = "https://github.com/joy-it/JoyIT_st7920/issues"